[REFACTOR] SearchPage 코드 분리
